refactor: move transaction types to a specific folder
