+++ src/app/(private)/transacoes/components/transaction-card.tsx
@@ -1,11 +1,11 @@
 "use client";
 
+import { TransactionType } from "@/src/app/(private)/transacoes/types/transaction-type";
 import Column from "@/src/components/utils/column";
 import Row from "@/src/components/utils/row";
 import Show from "@/src/components/utils/show";
 import CurrencyFormatter from "@/src/helpers/currency-formatter";
 import DateFormatter from "@/src/helpers/date-formatter";
-import { TransactionType } from "@/src/types/transaction-type";
 import TransactionDetailsDialog from "./transaction-details-dialog";
 
 interface TransactionCardProps {
